Modifications that lead to the firmware being able to build successfully

+++ src/modules/src/sysload.c
@@ -103,7 +103,7 @@ static void timerHandler(xTimerHandle timer) {
 
       uint32_t taskRunTime = stats->ulRunTimeCounter;
       float load = f * (taskRunTime - previousTaskData->ulRunTimeCounter);
-      DEBUG_PRINT("%.2f \t%u \t%s\n", (double)load, stats->usStackHighWaterMark, stats->pcTaskName);
+      DEBUG_PRINT("%.2f \t%u \t%s\n", (double)load, (unsigned int)stats->usStackHighWaterMark, stats->pcTaskName);
       
       previousTaskData->ulRunTimeCounter = taskRunTime;
     }
